nsfw/requirements.txt
numpy
tensorflow
tensorflow_hub
flask
Pillow
